refactor: test files import types from shared types layer
